Metrics: start the listeners only after the webhook ca is injected
The metrics serve as readiness probe. Helm wait checks for the pods to
be ready, and by doing this we ensure that the metric endpoint accepts
requests only after the webhook is ready to serve requests.
